Move StatusClaim and StatusListTokenClaim to Types.kt
