--- src/main/kotlin/eu/europa/ec/eudi/openid4vci/StatusClaim.kt
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2023-2026 European Commission
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package eu.europa.ec.eudi.openid4vci
-
-import eu.europa.ec.eudi.openid4vci.internal.URISerializer
-import kotlinx.serialization.Required
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-import java.net.URI
-
-@Serializable
-data class StatusClaim(
-    @Required @SerialName(TokenStatusListSpec.STATUS_LIST) val statusList: StatusListTokenClaim,
-)
-
-@Serializable
-data class StatusListTokenClaim(
-    @Required @SerialName(TokenStatusListSpec.INDEX) val index: UInt,
-    @Required @SerialName(TokenStatusListSpec.URI) @Serializable(with = URISerializer::class) val uri: URI,
-)
