+++ src/main/kotlin/eu/europa/ec/eudi/openid4vci/Types.kt
@@ -26,7 +26,10 @@ import com.nimbusds.jose.util.JSONObjectUtils
 import com.nimbusds.oauth2.sdk.`as`.ReadOnlyAuthorizationServerMetadata
 import eu.europa.ec.eudi.openid4vci.CredentialIssuanceError.ResponseEncryptionError.MissingRequiredRequestEncryptionSpecification
 import eu.europa.ec.eudi.openid4vci.internal.InstantEpochSecondSerializer
+import eu.europa.ec.eudi.openid4vci.internal.URISerializer
 import eu.europa.ec.eudi.openid4vci.internal.ensureNotNull
+import kotlinx.serialization.Required
+import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import java.net.URI
 import java.net.URL
@@ -397,3 +400,14 @@ value class JwtId(val value: String) {
 typealias InstantAsEpochSecond =
     @Serializable(with = InstantEpochSecondSerializer::class)
     Instant
+
+@Serializable
+data class StatusClaim(
+    @Required @SerialName(TokenStatusListSpec.STATUS_LIST) val statusList: StatusListTokenClaim,
+)
+
+@Serializable
+data class StatusListTokenClaim(
+    @Required @SerialName(TokenStatusListSpec.INDEX) val index: UInt,
+    @Required @SerialName(TokenStatusListSpec.URI) @Serializable(with = URISerializer::class) val uri: URI,
+)
